refactor: explicit promotion from enum flags

+++ app/Archive/abstractthreadedjob.cpp
@@ -194,7 +194,7 @@ void AbstractThreadedJob::slotDownloadResult(KJob *job)
         (*resultResp) << QVariant(job->error());
         (*resultResp) << QVariant(job->errorText());
     } else {
-        (*resultResp) << QVariant(KJob::UserDefinedError);
+        (*resultResp) << QVariant(static_cast<int>(KJob::UserDefinedError));
         (*resultResp) << QVariant(QString(i18n("Internal error, undefined <job> in result signal")));
     }
